api/requirements.txt
fastapi==0.99.1
websockets==11.0.3
uvicorn==0.23.2
python-multipart==0.0.9
fastapi_jwt_auth==0.5.0
pydantic==1.10.21
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
requests
qrcode[pil]
pyotp
pyjwt
bcrypt
pdf2image
pillow
pypdf
